61947: atarix83 feedback changes 19-04-2019
