+++ src/app/+search-page/search-filters/search-filter/search-filter.service.spec.ts
@@ -289,76 +289,4 @@ describe('SearchFilterService', () => {
     });
   });
 
-  describe('when the getPaginatedSearchOptions method is called', () => {
-    beforeEach(() => {
-      spyOn(service, 'getCurrentPagination');
-      spyOn(service, 'getCurrentSort');
-      spyOn(service, 'getCurrentView');
-      spyOn(service, 'getCurrentScope');
-      spyOn(service, 'getCurrentQuery');
-      spyOn(service, 'getCurrentFilters');
-      spyOn(service, 'getCurrentFixedFilter');
-      service.getPaginatedSearchOptions();
-    });
-
-    it('should call getCurrentPagination to build the paginated search options', () => {
-      expect(service.getCurrentPagination).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentSort to build the paginated search options', () => {
-      expect(service.getCurrentSort).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentView to build the paginated search options', () => {
-      expect(service.getCurrentView).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentScope to build the paginated search options', () => {
-      expect(service.getCurrentScope).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentQuery to build the paginated search options', () => {
-      expect(service.getCurrentQuery).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentFilters to build the paginated search options', () => {
-      expect(service.getCurrentFilters).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentFixedFilter to build the paginated search options', () => {
-      expect(service.getCurrentFixedFilter).toHaveBeenCalled();
-    });
-  });
-
-  describe('when the getSearchOptions method is called', () => {
-    beforeEach(() => {
-      spyOn(service, 'getCurrentView');
-      spyOn(service, 'getCurrentScope');
-      spyOn(service, 'getCurrentQuery');
-      spyOn(service, 'getCurrentFilters');
-      spyOn(service, 'getCurrentFixedFilter');
-      service.getPaginatedSearchOptions();
-    });
-
-    it('should call getCurrentView to build the search options', () => {
-      expect(service.getCurrentView).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentScope to build the search options', () => {
-      expect(service.getCurrentScope).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentQuery to build the search options', () => {
-      expect(service.getCurrentQuery).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentFilters to build the search options', () => {
-      expect(service.getCurrentFilters).toHaveBeenCalled();
-    });
-
-    it('should call getCurrentFixedFilter to build the search options', () => {
-      expect(service.getCurrentFixedFilter).toHaveBeenCalled();
-    });
-  });
-
 });
