+++ src/app/shared/object-list/item-list-element/item-types/journal-issue/journal-issue-list-element.component.ts
@@ -1,9 +1,8 @@
 import { Component } from '@angular/core';
-import { rendersItemType } from '../../../../items/item-type-decorator';
+import { ItemViewMode, rendersItemType } from '../../../../items/item-type-decorator';
 import { TypedItemSearchResultListElementComponent } from '../typed-item-search-result-list-element.component';
-import { VIEW_MODE_ELEMENT } from '../../../../../+item-page/simple/related-items/related-items-component';
 
-@rendersItemType('JournalIssue', VIEW_MODE_ELEMENT)
+@rendersItemType('JournalIssue', ItemViewMode.Element)
 @Component({
   selector: 'ds-journal-issue-list-element',
   styleUrls: ['./journal-issue-list-element.component.scss'],
