+++ src/app/shared/object-list/item-list-element/item-types/journal/journal-list-element.component.ts
@@ -1,9 +1,8 @@
 import { Component } from '@angular/core';
-import { rendersItemType } from '../../../../items/item-type-decorator';
+import { ItemViewMode, rendersItemType } from '../../../../items/item-type-decorator';
 import { TypedItemSearchResultListElementComponent } from '../typed-item-search-result-list-element.component';
-import { VIEW_MODE_ELEMENT } from '../../../../../+item-page/simple/related-items/related-items-component';
 
-@rendersItemType('Journal', VIEW_MODE_ELEMENT)
+@rendersItemType('Journal', ItemViewMode.Element)
 @Component({
   selector: 'ds-journal-list-element',
   styleUrls: ['./journal-list-element.component.scss'],
