+++ src/app/shared/object-list/item-list-element/item-types/publication/publication-list-element.component.ts
@@ -1,10 +1,9 @@
 import { Component } from '@angular/core';
-import { DEFAULT_ITEM_TYPE, rendersItemType } from '../../../../items/item-type-decorator';
+import { DEFAULT_ITEM_TYPE, ItemViewMode, rendersItemType } from '../../../../items/item-type-decorator';
 import { TypedItemSearchResultListElementComponent } from '../typed-item-search-result-list-element.component';
-import { VIEW_MODE_ELEMENT } from '../../../../../+item-page/simple/related-items/related-items-component';
 
-@rendersItemType('Publication', VIEW_MODE_ELEMENT)
-@rendersItemType(DEFAULT_ITEM_TYPE, VIEW_MODE_ELEMENT)
+@rendersItemType('Publication', ItemViewMode.Element)
+@rendersItemType(DEFAULT_ITEM_TYPE, ItemViewMode.Element)
 @Component({
   selector: 'ds-publication-list-element',
   styleUrls: ['./publication-list-element.component.scss'],
